+++ visualization/app/codeCharta/ui/codeMap/codeMap.arrow.service.spec.ts
@@ -22,6 +22,7 @@ import { setScaling } from "../../state/store/appSettings/scaling/scaling.action
 import { setEdges } from "../../state/store/fileSettings/edges/edges.actions"
 import { setHeightMetric } from "../../state/store/dynamicSettings/heightMetric/heightMetric.actions"
 import { CodeMapMesh } from "./rendering/codeMapMesh"
+import { setIsEdgeMetricVisible } from "../../state/store/appSettings/isEdgeMetricVisible/isEdgeMetricVisible.actions"
 
 describe("CodeMapArrowService", () => {
 	let codeMapArrowService: CodeMapArrowService
@@ -142,7 +143,7 @@ describe("CodeMapArrowService", () => {
 			expect(codeMapArrowService.addEdgePreview).toHaveBeenCalledTimes(0)
 		})
 
-		it("should call clearArrows and showEdgesOfBuildings through BuildingUnHovered", () => {
+		it("should call clearArrows and showEdgesOfBuildings through BuildingUnHovered when edge metric is enabled", () => {
 			codeMapArrowService.onBuildingUnhovered()
 
 			expect(codeMapArrowService.clearArrows).toHaveBeenCalled()
@@ -150,7 +151,16 @@ describe("CodeMapArrowService", () => {
 			expect(codeMapArrowService.addEdgePreview).toHaveBeenCalledTimes(0)
 		})
 
-		it("should call clearArrows and showEdgesOfBuildings through BuildingDeselcted", () => {
+		it("should call clearArrows and showEdgesOfBuildings through BuildingUnHovered when edge metric is disabled", () => {
+			storeService.dispatch(setIsEdgeMetricVisible())
+			codeMapArrowService.onBuildingUnhovered()
+
+			expect(codeMapArrowService.clearArrows).toHaveBeenCalledTimes(0)
+			expect(codeMapArrowService["showEdgesOfBuildings"]).toHaveBeenCalledTimes(0)
+			expect(codeMapArrowService.addEdgePreview).toHaveBeenCalledTimes(0)
+		})
+
+		it("should call clearArrows and showEdgesOfBuildings through BuildingDeselected", () => {
 			codeMapArrowService.onBuildingDeselected()
 
 			expect(codeMapArrowService.clearArrows).toHaveBeenCalled()